A little tweaking to bp::debug::attachDebugger() for OSX.  Be smart about whether we're currently being debugged, throw an alert to let developer know pid to attach to.

+++ src/libs/platform_utils/src/bpdebug.cpp
@@ -37,10 +37,54 @@
 #ifdef WIN32
 #include <Windows.h>
 #elif defined(MACOSX)
+#include <assert.h>
+#include <stdbool.h>
+#include <sys/types.h>
+#include <unistd.h>
+#include <sys/sysctl.h>
 #include <signal.h>
 #endif
 
 using namespace std;
+
+
+#ifdef MACOSX
+// from http://developer.apple.com/library/mac/#qa/qa2004/qa1361.html
+// Returns true if the current process is being debugged (either
+// running under the debugger or has a debugger attached post facto).
+static bool 
+AmIBeingDebugged(void)
+
+{
+    int                 junk;
+    int                 mib[4];
+    struct kinfo_proc   info;
+    size_t              size;
+
+    // Initialize the flags so that, if sysctl fails for some bizarre
+    // reason, we get a predictable result.
+
+    info.kp_proc.p_flag = 0;
+
+    // Initialize mib, which tells sysctl the info we want, in this case
+    // we're looking for information about a specific process ID.
+
+    mib[0] = CTL_KERN;
+    mib[1] = KERN_PROC;
+    mib[2] = KERN_PROC_PID;
+    mib[3] = getpid();
+
+    // Call sysctl.
+
+    size = sizeof(info);
+    junk = sysctl(mib, sizeof(mib) / sizeof(*mib), &info, &size, NULL, 0);
+    assert(junk == 0);
+
+    // We're being debugged if the P_TRACED flag is set.
+
+    return ( (info.kp_proc.p_flag & P_TRACED) != 0 );
+}
+#endif
 
 
 namespace bp {
@@ -54,13 +98,17 @@ void attachDebuggerImpl()
 #ifdef WIN32
     DebugBreak();
 #elif defined(MACOSX)
-    std::stringstream ss;
-    pid_t p = getpid();
-    ss << "Stopping process (attachDebugger() called) pid - " << p;
-    std::cerr << ss.str() << std::endl;
-    BPLOG_ERROR( ss.str() );    
-    kill(p, SIGSTOP);
-#else
+    if (AmIBeingDebugged()) {
+        __asm__ volatile ("int3");
+    } else {  
+        pid_t p = getpid();
+        std::stringstream ss;
+        ss << "Stopping process (attachDebugger() called) pid - " << p;
+        std::cerr << ss.str() << std::endl;
+        BPLOG_ERROR( ss.str() );    
+        showAlert(ss.str());
+        kill(p, SIGSTOP);
+    }
 #endif
 #endif
 }
